requirements.txt
fastapi
uvicorn
pyjwt
python-jose
ultralytics
sahi
opencv-python-headless 
pillow 
torch 
torchvision